requirements.txt
Flask>=2.0
cryptography>=38
Pillow>=9
